+++ action-sheet/src/index.ts
@@ -1,10 +1,10 @@
 import { registerPlugin } from '@capacitor/core';
 
 import type { ActionSheetPlugin } from './definitions';
-import { ActionSheetOptionStyle } from './definitions';
 
 const ActionSheet = registerPlugin<ActionSheetPlugin>('ActionSheet', {
   web: () => import('./web').then(m => new m.ActionSheetWeb()),
 });
 
-export { ActionSheet, ActionSheetOptionStyle };
+export * from './definitions';
+export { ActionSheet };