+++ app/src/definitions.ts
@@ -45,7 +45,7 @@ export interface AppState {
   isActive: boolean;
 }
 
-export interface AppUrlOpen {
+export interface URLOpenListenerEvent {
   /**
    * The URL the app was opened with.
    *
@@ -79,7 +79,7 @@ export interface AppLaunchUrl {
   url: string;
 }
 
-export interface AppRestoredResult {
+export interface RestoredListenerEvent {
   /**
    * The pluginId this result corresponds to. For example, `Camera`.
    *
@@ -114,6 +114,11 @@ export interface AppRestoredResult {
     message: string;
   };
 }
+
+export type StateChangeListener = (state: AppState) => void;
+export type URLOpenListener = (event: URLOpenListenerEvent) => void;
+export type RestoredListener = (event: RestoredListenerEvent) => void;
+export type BackButtonListener = () => void;
 
 export interface AppPlugin {
   /**
@@ -154,7 +159,7 @@ export interface AppPlugin {
    */
   addListener(
     eventName: 'appStateChange',
-    listenerFunc: (state: AppState) => void,
+    listenerFunc: StateChangeListener,
   ): PluginListenerHandle;
 
   /**
@@ -165,7 +170,7 @@ export interface AppPlugin {
    */
   addListener(
     eventName: 'appUrlOpen',
-    listenerFunc: (data: AppUrlOpen) => void,
+    listenerFunc: URLOpenListener,
   ): PluginListenerHandle;
 
   /**
@@ -173,11 +178,29 @@ export interface AppPlugin {
    * when an activity returns to an app that was closed, this call will return any data
    * the app was launched with, converted into the form of a result from a plugin call.
    *
+   * On Android, due to memory constraints on low-end devices, it's possible
+   * that, if your app launches a new activity, your app will be terminated by
+   * the operating system in order to reduce memory consumption.
+   *
+   * For example, that means the Camera API, which launches a new Activity to
+   * take a photo, may not be able to return data back to your app.
+   *
+   * To avoid this, Capacitor stores all restored activity results on launch.
+   * You should add a listener for `appRestoredResult` in order to handle any
+   * plugin call results that were delivered when your app was not running.
+   *
+   * Once you have that result (if any), you can update the UI to restore a
+   * logical experience for the user, such as navigating or selecting the
+   * proper tab.
+   *
+   * We recommend every Android app using plugins that rely on external
+   * Activities (for example, Camera) to have this event and process handled.
+   *
    * @since 1.0.0
    */
   addListener(
     eventName: 'appRestoredResult',
-    listenerFunc: (data: AppRestoredResult) => void,
+    listenerFunc: RestoredListener,
   ): PluginListenerHandle;
 
   /**
@@ -189,7 +212,7 @@ export interface AppPlugin {
    */
   addListener(
     eventName: 'backButton',
-    listenerFunc: () => void,
+    listenerFunc: BackButtonListener,
   ): PluginListenerHandle;
 
   /**
@@ -199,3 +222,15 @@ export interface AppPlugin {
    */
   removeAllListeners(): void;
 }
+
+/**
+ * @deprecated Use `RestoredListenerEvent`.
+ * @since 1.0.0
+ */
+export type AppRestoredResult = RestoredListenerEvent;
+
+/**
+ * @deprecated Use `URLOpenListenerEvent`.
+ * @since 1.0.0
+ */
+export type AppUrlOpen = URLOpenListenerEvent;