+++ browser/src/web.ts
@@ -1,6 +1,6 @@
 import { WebPlugin } from '@capacitor/core';
 
-import type { BrowserPlugin, BrowserOpenOptions } from './definitions';
+import type { BrowserPlugin, OpenOptions } from './definitions';
 
 export class BrowserWeb extends WebPlugin implements BrowserPlugin {
   _lastWindow: Window | null;
@@ -10,7 +10,7 @@ export class BrowserWeb extends WebPlugin implements BrowserPlugin {
     this._lastWindow = null;
   }
 
-  async open(options: BrowserOpenOptions): Promise<void> {
+  async open(options: OpenOptions): Promise<void> {
     this._lastWindow = window.open(options.url, options.windowName || '_blank');
   }