+++ haptics/src/web.ts
@@ -2,21 +2,21 @@ import { WebPlugin } from '@capacitor/core';
 
 import type {
   HapticsPlugin,
-  HapticsImpactOptions,
-  HapticsNotificationOptions,
+  ImpactOptions,
+  NotificationOptions,
   VibrateOptions,
 } from './definitions';
-import { HapticsImpactStyle, HapticsNotificationType } from './definitions';
+import { ImpactStyle, NotificationType } from './definitions';
 
 export class HapticsWeb extends WebPlugin implements HapticsPlugin {
   selectionStarted = false;
 
-  async impact(options?: HapticsImpactOptions): Promise<void> {
+  async impact(options?: ImpactOptions): Promise<void> {
     const pattern = this.patternForImpact(options?.style);
     this.vibrateWithPattern(pattern);
   }
 
-  async notification(options?: HapticsNotificationOptions): Promise<void> {
+  async notification(options?: NotificationOptions): Promise<void> {
     const pattern = this.patternForNotification(options?.type);
     this.vibrateWithPattern(pattern);
   }
@@ -40,23 +40,21 @@ export class HapticsWeb extends WebPlugin implements HapticsPlugin {
     this.selectionStarted = false;
   }
 
-  private patternForImpact(
-    style: HapticsImpactStyle = HapticsImpactStyle.Heavy,
-  ): number[] {
-    if (style === HapticsImpactStyle.Medium) {
+  private patternForImpact(style: ImpactStyle = ImpactStyle.Heavy): number[] {
+    if (style === ImpactStyle.Medium) {
       return [43];
-    } else if (style === HapticsImpactStyle.Light) {
+    } else if (style === ImpactStyle.Light) {
       return [20];
     }
     return [61];
   }
 
   private patternForNotification(
-    type: HapticsNotificationType = HapticsNotificationType.Success,
+    type: NotificationType = NotificationType.Success,
   ): number[] {
-    if (type === HapticsNotificationType.Warning) {
+    if (type === NotificationType.Warning) {
       return [30, 40, 30, 50, 60];
-    } else if (type === HapticsNotificationType.Error) {
+    } else if (type === NotificationType.Error) {
       return [27, 45, 50];
     }
     return [35, 65, 21];