+++ text-zoom/src/index.ts
@@ -6,4 +6,5 @@ const TextZoom = registerPlugin<TextZoomPlugin>('TextZoom', {
   ios: import('./ios').then(m => new m.TextZoomIOS()),
 });
 
+export * from './definitions';
 export { TextZoom };