+++ app-launcher/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/app-launcher@0.2.1...@capacitor/app-launcher@0.3.0) (2020-12-27)
+
+
+### Features
+
+* **app-launcher:** make openUrl open urls on web ([#152](https://github.com/ionic-team/capacitor-plugins/issues/152)) ([2eba7b4](https://github.com/ionic-team/capacitor-plugins/commit/2eba7b41a05be2f63c916f51aa4b36ef548ef87c))
+
+
+
+
+
+## [0.2.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/app-launcher@0.2.0...@capacitor/app-launcher@0.2.1) (2020-12-20)
+
+
+### Bug Fixes
+
+* support deprecated types from Capacitor 2 ([#139](https://github.com/ionic-team/capacitor-plugins/issues/139)) ([2d7127a](https://github.com/ionic-team/capacitor-plugins/commit/2d7127a488e26f0287951921a6db47c49d817336))
+
+
+
+
+
 # [0.2.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/app-launcher@0.1.0...@capacitor/app-launcher@0.2.0) (2020-12-02)
 
 **Note:** Version bump only for package @capacitor/app-launcher